fix: add spoofing of signingInfo
Some major Google Apps stopped relying on the deprecated
`packageInfo.signature` field and started using
`packageInfo.signingInfo` instead. This commit introduces spoofing for
the `signingInfo` field to maintain compatibility.

Unfortunately, spoofing the `signingInfo` field is a little more
involved than spoofing the `signature` field, because the former relies
on a `SigningDetails` class which is not part of the public SDK API [1].
Also, there is no public constructor for `SigningInfo` before SDK 35.
Therefore we rely on reflection to instantiate the `SigningDetails`
object, and to find the associated `SigningInfo` constructor.

The certificate string was converted from hexadecimal to base64
encoding. This change was made to include public key details in
`SigningInfo`, as Java requires PEM data to be loaded from a byte array.
The Android SDK provides a built-in base64 package for decoding
base64-encoded strings, which simplifies handling compared to
hexadecimal strings. Additionally, base64-encoded PEM certificates are
more commonly used.

See also:

- https://github.com/microg/GmsCore/issues/2680
- https://gitlab.e.foundation/e/os/android_frameworks_base/-/commit/2b9c74a4409f41924905c4b28aa900904e442992
- https://review.lineageos.org/c/LineageOS/android_frameworks_base/+/411374